Html template in progress for to do list
Advice required because it seems the html can only access files in the public folder. I have made a css sheet in the public folder for now since I can't figure out a way to reference the css files in the src folder. I also started writing the html for the to do list in the index.html which I'm not sure was the intended purpose for index.html.